requirements.txt
numpy==1.26.4
pandas==2.2.1

opencv-python-headless==4.9.0.80
albumentations==1.4.1
Pillow==10.2.0

segmentation-models-pytorch==0.3.3
timm==0.9.2

torchmetrics==1.3.1
scikit-learn==1.4.1.post1
scikit-image>=0.22.0        # SkeletonLoss 骨架预计算（tools/precompute_skeletons.py）
gudhi>=3.7.1                # TopologyLoss 持续同调计算

wandb==0.16.4
tensorboard==2.16.2

tqdm==4.66.2
matplotlib==3.8.3